file-based: move permissions transfer mode to general with abstract Identitie stream, so file based Identity stream becomes an implementation
